Move sqlite impl to it's own module
