+++ crates/locutus-node/src/ring.rs
@@ -46,11 +46,14 @@ pub(crate) struct Ring {
     /// of subscribers more often than inserting, and anyways is a relatively short sequence
     /// then is more optimal to just use a vector for it's compact memory layout.
     subscribers: DashMap<ContractKey, Vec<PeerKeyLocation>>,
+    subscriptions: Vec<PeerKeyLocation>,
 }
 
 impl Ring {
     const MIN_CONNECTIONS: usize = 10;
     const MAX_CONNECTIONS: usize = 20;
+    /// Max number of subscribers for a contract.
+    const MAX_SUBSCRIBERS: usize = 10;
 
     /// Above this number of remaining hops,
     /// randomize which of node a message which be forwarded to.
@@ -66,6 +69,7 @@ impl Ring {
             cached_contracts: DashSet::new(),
             own_location: RwLock::new(None),
             subscribers: DashMap::new(),
+            subscriptions: Vec::new(),
         }
     }
 
@@ -178,11 +182,25 @@ impl Ring {
             .copied()
     }
 
-    pub fn add_subscriber(&self, contract: ContractKey, subscriber: PeerKeyLocation) {
-        let mut subs = self.subscribers.entry(contract).or_default();
+    /// Will return an error in case the max number of subscribers has been added.
+    pub fn add_subscriber(
+        &self,
+        contract: ContractKey,
+        subscriber: PeerKeyLocation,
+    ) -> Result<(), ()> {
+        let mut subs = self
+            .subscribers
+            .entry(contract)
+            .or_insert(Vec::with_capacity(Self::MAX_SUBSCRIBERS));
         if let Err(next_idx) = subs.value_mut().binary_search(&subscriber) {
-            subs.value_mut().insert(next_idx, subscriber);
+            let subs = subs.value_mut();
+            if subs.len() == Self::MAX_SUBSCRIBERS {
+                return Err(());
+            } else {
+                subs.insert(next_idx, subscriber);
+            }
         }
+        Ok(())
     }
 }
